+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `Config::output_conflict_handling` does not contain the bless command message anymore, it is instead available separately as `Config::bless_command`
 * Updating `cargo_metadata` to `0.18`
 * Updated `spanned` to `0.1.5`, giving more precise spans for more iterator operations
+* `Config::cfgs` is now `Config::program::cfg_flag`
 
 ### Removed